separate history and matching server options
